Added ability to specify libraries needed by GLAs inside the comments of the GLA.
Also cleaned up some of the work functions such that they will generate code
that has correct indentation and is generally more easily read.

Corrected some issues with the Distinct GLA that I missed when I changed it
to be compatible with the changes to how local copies of data from chunks
must be made.