Initial implementation of Segmented Meta-GLA. Fixed small bugs in GroupBy and Distinct.
There were small issues in GroupBy and Distinct where they could produce empty
fragments. This should be fixed now.

There is an initial implementation of the Segmented Meta-GLA, which turns a
normal GLA into one with segmented state. It mostly works, but there are
some bizarre issues that cause aborts only sometimes. I'm currently
debugging this.

I also removed DataTypes.m4 in the Generated folder from version control, as it
was causing strange issues, as everything in the Generated folder is on
svn:ignore. Instead, a link to the correct file is automatically created if
it does not already exist.