Fixed issue with column access macros. Fixed issue with pattern matcher.
Fixed a problem with the column access macros where they could access the
current element and then immediately advance. Since GetCurrent now returns
a reference to an internal variable, the called to Advance would immediately
update the internal variable, which was not the intended behavior.
The iterators are now separately advanced after the values have been used.
This fixes an issue where HString values would sometimes be blank.

Fixed an issue with the pattern matcher so that the IsMatch method takes
a const HString instead of a non-const HString so that the compiler does
not freak out now that the values from the column iterators are const
references.

The change to TPCH-Q2 is some debugging statements.

KNOWN BUGS:
The join currently does not work properly. I'm assuming something went
wrong with the addition of the IN/NOT IN feature. It currently just
matches values that are equal to (N*40000 + 1) where N is a natural
number. If you want to check this out yourself, just run TPCH-Q2
and look at the output of j3.