Removed the cast to const char * operator from HString. Fixes to some GLAs.
Removal of the const char * operator from HString allows for the correct ==
operator to be automatically called without having to statically cast the
operands to HString. Before, the compiler was automatically converting both
operands to const char * and comparing them as pointers.

Fixed TopKGLA so that it is a macro again, not sure how it was uploaded
incorrectly.

Changed the GLUE_LISTS macro in GLA-Templates so that it will not do anything
if one of the lists is empty.