Revert revision 68 changes. Enhanced parser. Added armadillo example.
Reverted the use of anonymous namespaces, as they broke armadillo. I believe
this is because there are #include statements inside the anonymous namespace,
which causes all kinds of havoc with the naming.

Enhanced the parser so that datatypes, synonyms, functions, and GLAs can be
added to the datatype manager from Piggy.

Added an example of a GLA using armadillo (SparsityMatrixGLA). This GLA
computes the sparsity matrix of the incoming data set, with the input being
armadillo column vectors (created using a templated function).