Enhanced parser and type manager to support preprocessing, typedefs and more.
The type manager no longer stores information in a SQLite database, the
information about types, functions, and GLAs are read from libraries each
time a query is parsed.

All piggy files are now first run through the C Preprocessor, which allows for
includes, macros, and other nice things.

Note: As far as I can tell, the system is currently stable with the changes,
but the work on the parser is not quite finished. There is still some work
to be done before DataTypes.h is no longer needed, and the base library
is completely extracted from the system.