Enhanced Print statement.
For Print, the AS clause is now optional, and specifies the header to be
printed at the beginning of the output file. Each element in the AS clause
is a colon-separated list of identifiers. The first identifier in this list
will go on the first line, the second identifier on the second line, and so
on. The elements need not have the same number of items in their lists.
Each item will appear in the correct position on its line, with blank strings
inserted for elements that have no entry in that line.