Tweaked inheritance hierarchy of GLA and GT. Began adding GF capability to selection.
The GLA, GT, and Selection waypoints are now all "General Processing" waypoints.
The GPWayPointImp contains a framework that helps with tasks that are common
between the various types of processing these waypoints perform.

The Selection waypoint is being enhanced with the capability to use
Generalized Filters. Instead of a boolean expression being used to filter,
an object of a specified class can be used to perform the filtration. This
object, called a "Generalized Filter", can receive constant state(s) from
other waypoints to use during the filtration.