Updated documentation for GF, GT, GLA. Added some library checking to init scripts.
Added basic information about GFs, GTs, and GLAs to the tutorial, with some
simple examples.

The compilation scripts now check to ensure that required libraries are
installed using pkg-config. A README has been added explaining what
libraries are required by the system and what needs to be in place for the
system to determine that they are installed.