More improvements to profiling. Added thread-safe random functions.
Perf-based profiling can now be toggled on and off by changing the
PER_CPU_PROFILE define in the main makefile.

Per-query profiling in the work functions can be toggled by changing the
PER_QUERY_PROFILE define in the MK-Generated makefile.

Two random number functions, RandInt and RandDouble, have been added in a
new header, Random.h. These functions utilize thread-local storage for the
state of the random number generator, and are thus thread-safe without locking.
I have benchmarked these functions as being SIGNIFICANTLY faster than rand()
in multithreaded environments due to the spinlock in rand().

The RNG used is a Mersenne-Twister if C++11 libraries are available, otherwise
the rand48 class of functions are used (which are linear congruential). The
Mersenne-Twister RNG has a larger state than the linear congruential RNG and
is slower, but should produce better random numbers, and is more portable.