Fixed issue with counters in GIST. Fixed issue with GIST performance. Other minor fixes.
Changed all instances of 'realpath' to 'readlink -f' in scripts, as realpath
may not be available on all systems.

The GIST waypoint was not requesting tokens aggressively enough, this has been fixed.

Fixed minor issue in the TranslatorImp caused by an inclusion order problem. Because
ANTLR wraps all of the c++ includes inside an extern "C" environment, older compilers
will refuse to compile if a c++ header that defines templates is included inside
this environment and has not been included previously outside of the extern "C".